모임그레이드 추가, 와글추가 수정
