+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:supportsRtl="true"
         android:theme="@style/WAGLETheme"
         android:usesCleartextTraffic="true">
-
+        <activity android:name=".activity.wagleSub.AddWagleActivity"></activity>
         <activity android:name=".FindLocationActivity" />
         <activity android:name=".ViewDetailWagleActivity" />
         <activity android:name=".jhj.Jhj_Post_Gallery_List" />
@@ -32,7 +32,8 @@
         <activity android:name=".activity.wagleSub.AddCostActivity" />
         <activity android:name=".activity.wagleSub.AddDHGActivity" />
         <activity android:name=".activity.wagleSub.AddBJMActivity" />
-        <activity android:name=".activity.wagleSub.AddWagleActivity" />
+        <activity android:name=".activity.wagleSub.AddNormWagleActivity" />
+        <activity android:name=".activity.wagleSub.AddTodayWagleActivity" />
         <activity android:name=".MakeMoimActivity" />
         <activity android:name=".HomeActivity" />
         <activity android:name=".activity.user.SignUpActivity" />
@@ -40,12 +41,11 @@
         <activity android:name=".activity.user.LoginActivity" />
         <activity android:name=".MainActivity">
             <intent-filter>
-                <action android:name="android.intent.action.MAIN"/>
+                <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER"/>
+                <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
 
         <meta-data
             android:name="com.kakao.sdk.AppKey"
