+++ app/src/main/java/com/androidlec/wagle/fragments/HomeFragment.java
@@ -10,6 +10,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.ImageView;
+
 import com.androidlec.wagle.R;
 import com.androidlec.wagle.jhj.Jhj_FTPConnect;
 import com.androidlec.wagle.jhj.Jhj_Gallery_DTO;
@@ -89,6 +90,7 @@ public class HomeFragment extends Fragment {
         if (getArguments() != null) {
             mParam1 = getArguments().getString(ARG_PARAM1);
             mParam2 = getArguments().getString(ARG_PARAM2);
+
         }
     }
 
@@ -422,4 +424,5 @@ public class HomeFragment extends Fragment {
     // -------------------------------------------------------------------------------------
     // 갤러리 끝
     // -------------------------------------------------------------------------------------
+
 }
