+++ app/src/main/res/layout/fragment_waggle.xml
@@ -59,9 +59,11 @@
                 android:id="@+id/wagle_fab_addwagle"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_gravity="bottom|end"
+                android:layout_gravity="center|bottom"
+                android:textColor="@color/mainColor"
                 android:layout_margin="20dp"
-                android:background="@drawable/ic_add_circle_24" />
+                android:src="@drawable/ic_baseline_add_24"
+                 />
 
         </FrameLayout>
