+++ app/src/main/res/layout/include_actionbar.xml
@@ -9,7 +9,8 @@
         android:id="@+id/include_ab_iv"
         android:layout_width="?attr/actionBarSize"
         android:layout_height="?attr/actionBarSize"
-        android:background="@drawable/ic_outline_emptyimage"
+        android:padding="5dp"
+        android:src="@drawable/ic_outline_emptyimage"
         />
 
 
@@ -17,10 +18,11 @@
         android:id="@+id/include_ab_tv"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginLeft="20dp"
+        android:layout_marginLeft="24dp"
         android:layout_marginRight="4dp"
         android:text="Moim Name"
-        android:textSize="20dp">
+        android:textStyle="bold"
+        android:textSize="16dp">
     </TextView>
